LayoutTests/platform/ios-sim/iphone/touch/construct-TouchList.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../../fast/js/resources/js-test-style.css">
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
</script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script src="resources/construct-TouchList.js"></script>
<script src="../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
