LayoutTests/platform/ios-sim-deprecated/iphone/touch/construct-TouchList-expected.txt

CONSOLE MESSAGE: line 1: TypeError: HTMLParagraphElement is not a function (evaluating 'description(
'Test constructor for TouchList.  &lt;<a href="rdar://problem/6481690">rdar://problem/6481690</a>&gt;'
)')
